Refactor code: 1) the files of the films module have been created 2) the states films and serials have been changed Started work of serials module 1) the files of the serials module have been created 2) serials module is in development
