+++ tgbot/misc/states.py
@@ -3,4 +3,5 @@ from aiogram.dispatcher.filters.state import StatesGroup, State
 
 class UserState(StatesGroup):
     Category = State()
-    Film_name = State()
+    Films = State()
+    TV_Serials = State()
